Dialog

+++ app/src/containers/RightContainer.tsx
@@ -31,7 +31,7 @@ const RightContainer = ({ style }): JSX.Element => {
   const [BGColor, setBGColor] = useState('');
   const [compWidth, setCompWidth] = useState('');
   const [compHeight, setCompHeight] = useState('');
-  const [dialogError, setDialogError] = useState(false);
+  const [deleteLinkedPageError, setDeleteLinkedPageError] = useState(false);
   const [deleteIndexError, setDeleteIndexError] = useState(false);
   const [deleteComponentError, setDeleteComponentError] = useState(false);
 
@@ -150,8 +150,6 @@ const RightContainer = ({ style }): JSX.Element => {
   const isIndex = (): boolean => configTarget.id === 1;
 
   const isChildOfPage = (): boolean => {
-    // TODO: refactor
-    // TODO: output parent name and id to refocus canvas on parent
     let isChild: boolean = false;
     const { id } = configTarget;
     state.components.forEach(comp => {
@@ -162,7 +160,24 @@ const RightContainer = ({ style }): JSX.Element => {
       });
     });
     return isChild;
-  };
+  }
+
+  const isLinkedTo = (): boolean => {
+    const { id } = configTarget;
+    const pageName = state.components[id-1].name;
+    let isLinked = false;
+    const searchNestedChildren = (comps) => {
+      if (comps.length === 0) return;
+        comps.forEach((comp, i) => {
+          if (comp.type === 'Route Link' && comp.name === pageName) {
+            isLinked = true;
+          }
+          if (comp.children.length > 0) searchNestedChildren(comp.children);  
+        });
+      } 
+    searchNestedChildren(state.components);
+    return isLinked;
+  }
 
   // dispatch to 'UPDATE CSS' called when save button is clicked,
   // passing in style object constructed from all changed input values
@@ -189,13 +204,14 @@ const RightContainer = ({ style }): JSX.Element => {
     dispatch({ type: 'DELETE CHILD', payload: {} });
   };
 
-  const handlePageDelete = id => () => {
-    // TODO: return modal
-    isIndex()
-      ? handleDialogError('index')
-      : dispatch({ type: 'DELETE PAGE', payload: { id } });
-  };
-
+  const handlePageDelete = (id) => () => {
+    // TODO: return modal 
+    if (isLinkedTo()) return setDeleteLinkedPageError(true);
+    isIndex() 
+      ? handleDialogError('index') 
+      : dispatch({ type: 'DELETE PAGE', payload: { id }});
+  }
+  
   const handleDeleteReusableComponent = () => {
     isChildOfPage()
       ? handleDialogError('component')
@@ -216,7 +232,8 @@ const RightContainer = ({ style }): JSX.Element => {
   const handleCloseDialogError = () => {
     setDeleteIndexError(false);
     setDeleteComponentError(false);
-  };
+    setDeleteLinkedPageError(false);
+  }
 
   return (
     <div className="column right" style={style}>
@@ -443,17 +460,21 @@ const RightContainer = ({ style }): JSX.Element => {
         <ProjectManager />
       </div>
       <Dialog
-        open={deleteIndexError}
+        open={deleteIndexError || deleteLinkedPageError || deleteComponentError}
         onClose={handleCloseDialogError}
         aria-labelledby="alert-dialog-title"
         aria-describedby="alert-dialog-description"
       >
         <DialogTitle id="alert-dialog-title">
-          {ErrorMessages.deleteIndexTitle}
+          {deleteIndexError ? ErrorMessages.deleteIndexTitle : ''}
+          {deleteComponentError ? ErrorMessages.deleteComponentTitle : ''} 
+          {deleteLinkedPageError ? ErrorMessages.deleteLinkedPageTitle : ''}
         </DialogTitle>
         <DialogContent>
           <DialogContentText id="alert-dialog-description">
-            {ErrorMessages.deleteIndexMessage}
+          {deleteIndexError ? ErrorMessages.deleteIndexMessage : ''}
+          {deleteComponentError ? ErrorMessages.deleteComponentMessage : ''} 
+          {deleteLinkedPageError ? ErrorMessages.deleteLinkedPageMessage : ''}
           </DialogContentText>
         </DialogContent>
         <DialogActions>
@@ -462,7 +483,7 @@ const RightContainer = ({ style }): JSX.Element => {
           </Button>
         </DialogActions>
       </Dialog>
-      <Dialog
+      {/* <Dialog
         open={deleteComponentError}
         onClose={handleCloseDialogError}
         aria-labelledby="alert-dialog-title"
@@ -481,7 +502,7 @@ const RightContainer = ({ style }): JSX.Element => {
             OK
           </Button>
         </DialogActions>
-      </Dialog>
+      </Dialog> */}
     </div>
   );
 };
